fix: move api compatibility types into namespace
