+++ templates/Sekiban.Dcb.Templates/content/Sekiban.Dcb.Orleans.Decider/SekibanDcbDecider.ApiService/OrleansCompatibilityTypes.cs
@@ -1,0 +1,49 @@
+using Orleans.Runtime;
+using Orleans.Runtime.Hosting;
+using Orleans.Storage;
+
+namespace SekibanDcbDecider.ApiService;
+
+sealed class CompatibleMemoryGrainStorage : IGrainStorage, ILifecycleParticipant<ISiloLifecycle>, IDisposable
+{
+    private readonly MemoryGrainStorage _inner;
+
+    public CompatibleMemoryGrainStorage(MemoryGrainStorage inner)
+    {
+        _inner = inner;
+    }
+
+    Task IGrainStorage.ReadStateAsync<T>(
+        string stateName,
+        GrainId grainId,
+        IGrainState<T> grainState) =>
+        _inner.ReadStateAsync(stateName, grainId, grainState);
+
+    Task IGrainStorage.WriteStateAsync<T>(
+        string stateName,
+        GrainId grainId,
+        IGrainState<T> grainState) =>
+        _inner.WriteStateAsync(stateName, grainId, grainState);
+
+    Task IGrainStorage.ClearStateAsync<T>(
+        string stateName,
+        GrainId grainId,
+        IGrainState<T> grainState) =>
+        _inner.ClearStateAsync(stateName, grainId, grainState);
+
+    public void Participate(ISiloLifecycle lifecycle)
+    {
+        // MemoryGrainStorageFactory already registers the lifecycle behavior for the
+        // wrapped storage instance, so this adapter intentionally does nothing here.
+    }
+
+    public void Dispose()
+    {
+        _inner.Dispose();
+    }
+}
+
+readonly record struct PostgresBootstrapSettings(
+    string DatabaseName,
+    string EscapedDatabaseName,
+    string AdminConnectionString);
